[CL] Extract create reaction rate.
